+++ packages/common/src/initiatives/HubInitiatives.ts
@@ -13,6 +13,7 @@ import {
   getUniqueSlug,
   setSlugKeyword,
 } from "../items/slugs";
+
 import {
   isGuid,
   cloneObject,
@@ -237,9 +238,15 @@ export async function enrichInitiativeSearchResult(
   // TODO: Link handling should be an enrichment
   result.links.thumbnail = getItemThumbnailUrl(item, requestOptions);
   result.links.self = getItemHomeUrl(result.id, requestOptions);
+  const identifier = item.id;
+  // Until the new initiatives route is in place we need to
+  // route using the id. Once the route is in place we can
+  // un-comment this
+  // const identifier = getItemIdentifier(item);
+
   result.links.siteRelative = getHubRelativeUrl(
     result.type,
-    result.id,
+    identifier,
     item.typeKeywords
   );