+++ packages/common/src/projects/_internal/ProjectUiSchemaCreate.ts
@@ -76,6 +76,11 @@ export const uiSchema: IUiSchema = {
             {
               type: "Section",
               labelKey: "{{i18nScope}}.sections.location.label",
+              options: {
+                helperText: {
+                  labelKey: "{{i18nScope}}.sections.location.helperText",
+                },
+              },
               elements: [
                 {
                   scope: "/properties/location",