+++ packages/discussions/src/utils/channels/index.ts
@@ -1,6 +1,6 @@
 import { IUser } from "@esri/arcgis-rest-auth";
 import { GroupMembership } from "@esri/arcgis-rest-portal";
-import { IChannel, IPlatformSharing } from "../../types";
+import { IChannel, IDiscussionsUser, IPlatformSharing } from "../../types";
 import { isOrgAdmin, reduceByGroupMembership } from "../platform";
 
 export { canPostToChannel } from "./can-post-to-channel";
@@ -9,9 +9,9 @@ export { canCreateChannel } from "./can-create-channel";
 function intersectGroups(
   membershipTypes: GroupMembership[]
 ): (arg0: IUser, arg1: IChannel) => boolean {
-  return (user: IUser, channel: IChannel): boolean => {
-    const { groups: sharedGroups } = channel;
-    const { groups: userGroups } = user;
+  return (user: IUser | IDiscussionsUser, channel: IChannel): boolean => {
+    const { groups: sharedGroups = [] } = channel;
+    const { groups: userGroups = [] } = user;
     const eligibleUserGroups = userGroups.reduce(
       reduceByGroupMembership(membershipTypes),
       []
@@ -23,12 +23,18 @@ function intersectGroups(
   };
 }
 
-function isChannelOrgMember(channel: IChannel, user: IUser): boolean {
+function isChannelOrgMember(
+  channel: IChannel,
+  user: IUser | IDiscussionsUser
+): boolean {
   // orgs.length = 1 until collaboration/discussion between many orgs is ideated
   return channel.orgs.length === 1 && channel.orgs.indexOf(user.orgId) > -1;
 }
 
-function isChannelOrgAdmin(channel: IChannel, user: IUser): boolean {
+function isChannelOrgAdmin(
+  channel: IChannel,
+  user: IUser | IDiscussionsUser
+): boolean {
   return isOrgAdmin(user) && channel.orgs.indexOf(user.orgId) > -1;
 }
 
@@ -40,13 +46,23 @@ function isChannelOrgAdmin(channel: IChannel, user: IUser): boolean {
  * @param {IUser} user
  * @return {*}  {boolean}
  */
-export function canReadFromChannel(channel: IChannel, user: IUser): boolean {
+export function canReadFromChannel(
+  channel: IChannel,
+  user: IUser | IDiscussionsUser = {}
+): boolean {
   if (channel.access === "private") {
     // ensure user is member of at least one group
     return intersectGroups(["member", "owner", "admin"])(user, channel);
-  } else if (channel.access === "org") {
-    return isChannelOrgMember(channel, user);
   }
+
+  if (channel.access === "org") {
+    return (
+      intersectGroups(["member", "owner", "admin"])(user, channel) ||
+      isChannelOrgMember(channel, user)
+    );
+  }
+
+  // public channel
   return true;
 }
 
@@ -58,7 +74,14 @@ export function canReadFromChannel(channel: IChannel, user: IUser): boolean {
  * @param {IUser} user
  * @return {*}  {boolean}
  */
-export function canModifyChannel(channel: IChannel, user: IUser): boolean {
+export function canModifyChannel(
+  channel: IChannel,
+  user: IUser | IDiscussionsUser = {}
+): boolean {
+  if (!user.username) {
+    return false;
+  }
+
   if (channel.creator === user.username) {
     return true;
   }
@@ -67,8 +90,12 @@ export function canModifyChannel(channel: IChannel, user: IUser): boolean {
     // ensure user is owner/admin of at least one group
     return intersectGroups(["owner", "admin"])(user, channel);
   }
-  // if org or public channel, must be org admin
-  return isChannelOrgAdmin(channel, user);
+
+  // org or public channel
+  return (
+    intersectGroups(["owner", "admin"])(user, channel) ||
+    isChannelOrgAdmin(channel, user)
+  );
 }
 
 /**