+++ packages/discussions/test/utils/posts/index.test.ts
@@ -1,5 +1,10 @@
 import { IGroup, IUser } from "@esri/arcgis-rest-types";
-import { IChannel, IDiscussionParams, IPost } from "../../../src/types";
+import {
+  IChannel,
+  IDiscussionParams,
+  IDiscussionsUser,
+  IPost,
+} from "../../../src/types";
 import {
   isDiscussable,
   parseDiscussionURI,
@@ -103,7 +108,7 @@ describe("canDeletePost", () => {
     expect(canModifyChannelSpy).not.toHaveBeenCalled();
   });
 
-  it("returns when user can modify channel", () => {
+  it("returns true when user did not create the post and user can modify channel", () => {
     const canModifyChannelSpy = spyOn(
       channelUtils,
       "canModifyChannel"
@@ -129,6 +134,33 @@ describe("canDeletePost", () => {
     expect(result).toBe(false);
     expect(canModifyChannelSpy).toHaveBeenCalledTimes(1);
     expect(canModifyChannelSpy).toHaveBeenCalledWith(channel, user);
+  });
+
+  it("returns false when the user is unauthenticated", () => {
+    const canModifyChannelSpy = spyOn(
+      channelUtils,
+      "canModifyChannel"
+    ).and.returnValue(false);
+    const post = { id: "post1", creator: "user1" } as IPost;
+    const user = { username: null } as IDiscussionsUser;
+    const channel = { id: "channel1" } as IChannel;
+    const result = canDeletePost(post, channel, user);
+    expect(result).toBe(false);
+    expect(canModifyChannelSpy).toHaveBeenCalledTimes(1);
+    expect(canModifyChannelSpy).toHaveBeenCalledWith(channel, user);
+  });
+
+  it("returns false when the user is undefined", () => {
+    const canModifyChannelSpy = spyOn(
+      channelUtils,
+      "canModifyChannel"
+    ).and.returnValue(false);
+    const post = { id: "post1", creator: "user1" } as IPost;
+    const channel = { id: "channel1" } as IChannel;
+    const result = canDeletePost(post, channel);
+    expect(result).toBe(false);
+    expect(canModifyChannelSpy).toHaveBeenCalledTimes(1);
+    expect(canModifyChannelSpy).toHaveBeenCalledWith(channel, {});
   });
 });